Move telemetry provider to its own file
